chore: use mrdox::ToolExecutor
